docs/examples/file-browser/requirements.txt
shiny-treeview